refactor: improve structure of petInteractions component
