+++ frontend/src/components/pet/PetInteractions.css
@@ -20,6 +20,80 @@
     margin: 0;
     width: 100%;
     box-sizing: border-box;
+
+    .picker {
+        display: grid;
+        grid-template-columns: 1fr 1fr;
+        gap: 20px;
+        margin-bottom: 20px;
+
+        .field {
+            display: flex;
+            flex-direction: column;
+            gap: 8px;
+
+            label {
+                font-weight: 600;
+                color: var(--zi-text);
+                font-size: 14px;
+            }
+
+            select {
+                padding: 12px 16px;
+                border: 2px solid var(--zi-border);
+                border-radius: 8px;
+                font-size: 14px;
+                background: var(--zi-card-bg);
+                color: var(--zi-text);
+                cursor: pointer;
+                transition: all 0.3s ease;
+            }
+
+            select:focus {
+                outline: none;
+                border-color: #667eea;
+                box-shadow: 0 0 0 3px rgba(102, 126, 234, 0.1);
+            }
+
+            input {
+                padding: 12px 16px;
+                border: 2px solid var(--zi-border);
+                border-radius: 8px;
+                font-size: 14px;
+                background: var(--zi-card-bg);
+                color: var(--zi-text);
+                transition: all 0.3s ease;
+            }
+
+            input:focus {
+                outline: none;
+                border-color: #667eea;
+                box-shadow: 0 0 0 3px rgba(102, 126, 234, 0.1);
+            }
+        }
+    }
+
+    .interface {
+        background: var(--zi-surface-bg);
+        border-radius: 12px;
+        padding: 20px;
+        margin-top: 20px;
+
+        h4 {
+            margin: 0 0 8px 0;
+            font-size: 20px;
+            font-weight: 700;
+            color: var(--zi-text);
+            text-align: center;
+        }
+
+        p {
+            margin: 0 0 20px 0;
+            color: var(--zi-text-muted);
+            text-align: center;
+            font-size: 14px;
+        }
+    }
 }
 
 /* Full-page /breed, /battle, /levelup, /rename — no gallery */
@@ -359,67 +433,6 @@
     box-shadow: none;
 }
 
-.breed-interface,
-.battle-interface {
-    background: var(--zi-surface-bg);
-    border-radius: 12px;
-    padding: 20px;
-    margin-top: 20px;
-}
-
-.breed-interface h4,
-.battle-interface h4 {
-    margin: 0 0 8px 0;
-    font-size: 20px;
-    font-weight: 700;
-    color: var(--zi-text);
-    text-align: center;
-}
-
-.breed-interface p,
-.battle-interface p {
-    margin: 0 0 20px 0;
-    color: var(--zi-text-muted);
-    text-align: center;
-    font-size: 14px;
-}
-
-.pet-selection {
-    display: grid;
-    grid-template-columns: 1fr 1fr;
-    gap: 20px;
-    margin-bottom: 20px;
-}
-
-.selection-group {
-    display: flex;
-    flex-direction: column;
-    gap: 8px;
-}
-
-.selection-group label {
-    font-weight: 600;
-    color: var(--zi-text);
-    font-size: 14px;
-}
-
-.selection-group select {
-    padding: 12px 16px;
-    border: 2px solid var(--zi-border);
-    border-radius: 8px;
-    font-size: 14px;
-    background: var(--zi-card-bg);
-    color: var(--zi-text);
-    cursor: pointer;
-    transition: all 0.3s ease;
-}
-
-.selection-group select:focus {
-    outline: none;
-    border-color: #667eea;
-    box-shadow: 0 0 0 3px rgba(102, 126, 234, 0.1);
-}
-
 .name-input {
     display: flex;
     flex-direction: column;
@@ -573,7 +586,7 @@
         grid-template-columns: 1fr;
     }
 
-    .pet-selection {
+    .pet-interactions .picker {
         grid-template-columns: 1fr;
         gap: 16px;
     }
@@ -594,8 +607,7 @@
         font-size: 20px;
     }
 
-    .breed-interface h4,
-    .battle-interface h4 {
+    .pet-interactions .interface h4 {
         font-size: 18px;
     }
 }
